build shell script smartened up

+++ build.sh
@@ -1,3 +1,20 @@
+echo Usage: build.sh [version]
+
+if [ $# -eq 0 ]
+then
+    VERSION="latest"
+else
+    VERSION=$1
+fi
+
+NAME="flock-$VERSION"
+NAMEMIN="flock-$VERSION-min"
+
+if [ ! -d build ]
+then
+    mkdir build
+fi
+
 cat \
 src/license.js \
 src/flock.js \
@@ -7,10 +24,15 @@ src/live.js \
 src/event.js \
 src/constants.js \
 src/query.js \
-> build/flock-latest.js
+> "build/$NAME.js"
 
-jsmin < build/flock-latest.js > build/tmp.js
+if type jsmin >/dev/null 2>&1
+then
+    jsmin < "build/$NAME.js" > build/tmp.js
+    cat src/license.js build/tmp.js > "build/$NAMEMIN.js"
+    rm build/tmp.js
+else
+    echo Jsmin not found. Skipping minification.
+fi
 
-cat src/license.js build/tmp.js > build/flock-latest-min.js
-
-rm build/tmp.js
+echo Done.
